+++ src/components/code-snippet-title/code-snippet-title.js
@@ -42,7 +42,7 @@ export default class CodeSnippetTitle extends React.PureComponent {
 CodeSnippetTitle.propTypes = {
   /* Filename to be displayed as a kind of title. Can be a specific filename like
   `RuntimeStylingActivity.java` or a general filename like `Activity`. */
-  filename: PropTypes.string.isRequired,
+  filename: PropTypes.string,
   /* Optional `link` to a GitHub file. If this is set, the rendered component
   will include a "View on Github" link. */
   link: function (props, propName, componentName) {