+++ src/components/related-page/vimeo.js
@@ -93,11 +93,10 @@ export class VimeoPlayImage extends React.PureComponent {
           viewBox="0 0 20 20"
           width="100%"
           focusable="false"
-          aria-labelledby="play-icon-title"
+          aria-label="Play"
           role="img"
           fill={this.props.fallbackIcon ? '#7753eb' : '#fff'}
         >
-          <title id="play-icon-title">Play</title>
           <polygon points="1,0 20,10 1,20" />
         </svg>
       </IconWrapper>