+++ CHANGELOG.md
@@ -13,6 +13,8 @@ hideFromSearchEngines: true
 ---*/
 ```
 
+- Replace `id` with `aria-label` in `VimeoPlayImage` to prevent duplicated `id`s when more than one instance is on the same page. [#438](https://github.com/mapbox/dr-ui/pull/438)
+- Make `filename` optional for `CodeSnippetTitle`.
 - 🚨The `CodeSnippet` prop `highlighter` no longer accepts an arrow function.
   - Before: `highlighter={() => highlightJson}`
   - After: `highlighter={highlightJson}`